backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
openai==1.3.7
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
pydantic==2.5.0
email-validator==2.1.0.post1
alembic==1.12.1
pgvector==0.2.4


